move ImageMerkleTree and ImageOracle into core
