+++ waldo/core/Cargo.toml
@@ -6,12 +6,14 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bincode = "1.3"
 bytemuck = "1.12"
+divrem = "1.0.0"
+image = { version = "0.24.5", default-features = false }
 merkle_light = "0.4.0"
 risc0-zkp = { version = "1.0.0-rc.2", default-features = false }
 risc0-zkvm = { version = "1.0.0-rc.2", default-features = false, features = ["std", "metal"] }
 serde = "1.0"
-bincode = "1.3"
 
 [dev-dependencies]
 hex = "0.4.3"
