+++ waldo/methods/guest/Cargo.toml
@@ -9,8 +9,6 @@ edition = "2021"
 risc0-build = "1.0.0-rc.2"
 
 [dependencies]
-bytemuck = "1.12"
-divrem = "1.0.0"
 image = { version = "0.24.5", default-features = false }
 merkle_light = "0.4"
 once_cell = "1.17"
